Moved File System utilities/helper package from in_red_cfs/fs to fs.
