+++ fs/erasure/encoder.go
