+++ in_red_cfs/integration_tests/basic_ec_test.go
@@ -8,7 +8,7 @@ import (
 
 	"github.com/SharedCode/sop"
 	"github.com/SharedCode/sop/in_red_cfs"
-	"github.com/SharedCode/sop/in_red_cfs/fs"
+	"github.com/SharedCode/sop/fs"
 )
 
 func init() {
