+++ in_red_cfs/transaction.go
@@ -6,7 +6,7 @@ import (
 
 	"github.com/SharedCode/sop"
 	cas "github.com/SharedCode/sop/cassandra"
-	"github.com/SharedCode/sop/in_red_cfs/fs"
+	"github.com/SharedCode/sop/fs"
 	"github.com/SharedCode/sop/in_red_ck"
 )
